refactor: nest slide components under deck (#363)
* fix: remove duplicate deck utilities

* chore: restructure component directory layout and update imports

* fix: update Slide import path in renderDirectiveMarkdown.ts

* chore: rename Deck component file and update import paths

---------
