+++ apps/campfire/src/components/Campfire/__tests__/Campfire.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, act, waitFor } from '@testing-library/preact'
 import { Campfire } from '@campfire/components/Campfire'
-import { useStoryDataStore } from '@campfire/use-story-data-store'
-import { useGameStore } from '@campfire/use-game-store'
+import { useStoryDataStore } from '@campfire/state/useStoryDataStore'
+import { useGameStore } from '@campfire/state/useGameStore'
 import { samplePassage } from '@campfire/test-utils/helpers'
 import { describe, it, expect, beforeEach } from 'bun:test'
 import i18next from 'i18next'
