+++ apps/campfire/src/components/DebugWindow/__tests__/DebugWindow.test.tsx
@@ -1,8 +1,8 @@
 import { describe, it, expect, beforeEach, mock } from 'bun:test'
 import { render, screen, act } from '@testing-library/preact'
 import { DebugWindow } from '@campfire/components/DebugWindow'
-import { useStoryDataStore } from '@campfire/use-story-data-store'
-import { useGameStore } from '@campfire/use-game-store'
+import { useStoryDataStore } from '@campfire/state/useStoryDataStore'
+import { useGameStore } from '@campfire/state/useGameStore'
 import i18next from 'i18next'
 import type { Element } from 'hast'
