+++ apps/campfire/src/components/DebugWindow/index.tsx
@@ -2,8 +2,8 @@ import { useState, useEffect, useRef } from 'preact/hooks'
 import {
   useStoryDataStore,
   type StoryDataState
-} from '@campfire/use-story-data-store'
-import { useGameStore } from '@campfire/use-game-store'
+} from '@campfire/state/useStoryDataStore'
+import { useGameStore } from '@campfire/state/useGameStore'
 import i18next from 'i18next'
 import type { Element, Text, Content } from 'hast'
