+++ apps/campfire/src/components/Deck/Slide/__tests__/Slide.test.tsx
@@ -1,8 +1,8 @@
 import { describe, it, expect, beforeEach } from 'bun:test'
 import { render, screen } from '@testing-library/preact'
 import { Deck } from '@campfire/components/Deck'
-import { Slide, type SlideProps } from '@campfire/components/Slide'
-import { useDeckStore } from '@campfire/use-deck-store'
+import { Slide, type SlideProps } from '@campfire/components/Deck/Slide'
+import { useDeckStore } from '@campfire/state/useDeckStory'
 import type { VNode } from 'preact'
 import { StubAnimation } from '@campfire/test-utils/stub-animation'
