+++ apps/campfire/src/components/Deck/Slide/__tests__/SlideDirective.test.tsx
@@ -1,8 +1,8 @@
 import { describe, it, expect, beforeEach } from 'bun:test'
 import { render, act, fireEvent } from '@testing-library/preact'
 import { Deck, Slide, TriggerButton } from '@campfire/components'
-import { useDeckStore } from '@campfire/use-deck-store'
-import { useGameStore } from '@campfire/use-game-store'
+import { useDeckStore } from '@campfire/state/useDeckStory'
+import { useGameStore } from '@campfire/state/useGameStore'
 import { resetStores } from '@campfire/test-utils/helpers'
 import { StubAnimation } from '@campfire/test-utils/stub-animation'
 import { unified } from 'unified'
