+++ apps/campfire/src/components/Deck/Slide/__tests__/renderDirectiveMarkdown.test.tsx
@@ -1,8 +1,8 @@
 import { describe, it, expect, beforeEach } from 'bun:test'
 import { render } from '@testing-library/preact'
-import { renderDirectiveMarkdown } from '@campfire/components/Slide/renderDirectiveMarkdown'
+import { renderDirectiveMarkdown } from '@campfire/components/Deck/Slide/renderDirectiveMarkdown'
 import { useDirectiveHandlers } from '@campfire/hooks/useDirectiveHandlers'
-import { useGameStore } from '@campfire/use-game-store'
+import { useGameStore } from '@campfire/state/useGameStore'
 import { resetStores } from '@campfire/test-utils/helpers'
 import type { ComponentChild } from 'preact'
