+++ apps/campfire/src/components/Deck/Slide/renderDirectiveMarkdown.ts
@@ -14,8 +14,8 @@ import { TriggerButton } from '@campfire/components/Passage/TriggerButton'
 import { If } from '@campfire/components/Passage/If'
 import { Show } from '@campfire/components/Passage/Show'
 import { OnExit } from '@campfire/components/Passage/OnExit'
-import { Deck } from '@campfire/components/Deck/Deck'
-import { Slide } from '@campfire/components/Slide/Slide'
+import { Deck } from '@campfire/components/Deck'
+import { Slide } from './'
 
 /**
  * Converts Markdown containing Campfire directives into Preact elements.
