+++ apps/campfire/src/components/Deck/__tests__/Deck.test.tsx
@@ -1,8 +1,8 @@
 import { describe, it, expect, beforeEach } from 'bun:test'
 import { render, screen, fireEvent, act } from '@testing-library/preact'
 import { Deck } from '@campfire/components/Deck'
-import { Slide } from '@campfire/components/Slide'
-import { useDeckStore } from '@campfire/use-deck-store'
+import { Slide } from '@campfire/components/Deck/Slide'
+import { useDeckStore } from '@campfire/state/useDeckStory'
 import { StubAnimation } from '@campfire/test-utils/stub-animation'
 
 /**
