+++ apps/campfire/src/components/Passage/If.tsx
@@ -8,7 +8,7 @@ import remarkRehype from 'remark-rehype'
 import rehypeCampfire from '@campfire/rehype-campfire'
 import rehypeReact from 'rehype-react'
 import type { RootContent, Root } from 'mdast'
-import { useGameStore } from '@campfire/use-game-store'
+import { useGameStore } from '@campfire/state/useGameStore'
 import { useDirectiveHandlers } from '@campfire/hooks/useDirectiveHandlers'
 import { LinkButton } from '@campfire/components/Passage/LinkButton'
 import { TriggerButton } from '@campfire/components/Passage/TriggerButton'
