+++ apps/campfire/src/components/Passage/__tests__/Passage.lifecycle.test.tsx
@@ -5,8 +5,8 @@ import { initReactI18next } from 'react-i18next'
 import type { Element } from 'hast'
 import { Passage } from '@campfire/components/Passage/Passage'
 import { OnExit } from '@campfire/components/Passage/OnExit'
-import { useStoryDataStore } from '@campfire/use-story-data-store'
-import { useGameStore } from '@campfire/use-game-store'
+import { useStoryDataStore } from '@campfire/state/useStoryDataStore'
+import { useGameStore } from '@campfire/state/useGameStore'
 import { resetStores } from '@campfire/test-utils/helpers'
 import { unified } from 'unified'
 import remarkParse from 'remark-parse'
