+++ apps/campfire/src/components/Passage/__tests__/Passage.state.test.tsx
@@ -4,8 +4,8 @@ import i18next from 'i18next'
 import { initReactI18next } from 'react-i18next'
 import type { Element } from 'hast'
 import { Passage } from '@campfire/components/Passage/Passage'
-import { useStoryDataStore } from '@campfire/use-story-data-store'
-import { useGameStore } from '@campfire/use-game-store'
+import { useStoryDataStore } from '@campfire/state/useStoryDataStore'
+import { useGameStore } from '@campfire/state/useGameStore'
 import { resetStores } from '@campfire/test-utils/helpers'
 
 describe('Passage game state directives', () => {
