+++ apps/campfire/src/components/index.ts
@@ -9,8 +9,8 @@ export { Show } from '@campfire/components/Passage/Show'
 export { OnExit } from '@campfire/components/Passage/OnExit'
 export { DebugWindow } from '@campfire/components/DebugWindow'
 export { Deck } from '@campfire/components/Deck'
-export { Slide } from '@campfire/components/Slide'
-export { renderDirectiveMarkdown } from '@campfire/components/Slide/renderDirectiveMarkdown'
-export { Appear } from '@campfire/components/Appear'
-export { Layer } from '@campfire/components/Layer'
-export { Text } from '@campfire/components/Text'
+export { Slide } from '@campfire/components/Deck/Slide'
+export { renderDirectiveMarkdown } from '@campfire/components/Deck/Slide/renderDirectiveMarkdown'
+export { Appear } from '@campfire/components/Deck/Slide/Appear'
+export { Layer } from '@campfire/components/Deck/Slide/Layer'
+export { Text } from '@campfire/components/Deck/Slide/Text'
