+++ apps/campfire/src/hooks/__tests__/deckDirective.test.tsx
@@ -11,8 +11,8 @@ import rehypeCampfire from '@campfire/rehype-campfire'
 import rehypeReact from 'rehype-react'
 import type { ComponentChild } from 'preact'
 import { useDirectiveHandlers } from '@campfire/hooks/useDirectiveHandlers'
-import { Deck } from '@campfire/components/Deck/Deck'
-import { Slide } from '@campfire/components/Slide/Slide'
+import { Deck } from '@campfire/components/Deck'
+import { Slide } from '@campfire/components/Deck/Slide'
 import { LinkButton } from '@campfire/components/Passage/LinkButton'
 import { TriggerButton } from '@campfire/components/Passage/TriggerButton'
 import { If } from '@campfire/components/Passage/If'
