+++ apps/campfire/src/state/stateManager.ts
@@ -1,6 +1,6 @@
 import rfdc from 'rfdc'
 import { clamp } from '@campfire/remark-campfire/helpers'
-import { useGameStore } from '@campfire/use-game-store'
+import { useGameStore } from '@campfire/state/useGameStore'
 
 /** Range value representation */
 export interface RangeValue {
